+++ CHANGELOG.md
@@ -153,6 +153,7 @@ Compatibility:
 * Update `IO.readlines` to ignore negative limit parameters (#2625 , @bjfish).
 * Update `Math.sqrt` to raise a `Math::DomainError` for negative numbers (#2621, @bjfish).
 * Update `Enumerable#inject` to raise an `ArgumentError` if no block or symbol are given (#2626, @bjfish).
+* Update `Module#{include,prepend}` to affect previous includes/prepends (@bjfish, #2453)
 
 Performance: